dumpsd: move basic decode classes (atom, aggie) to separate file
