+++ tools/utils/dumpsd/dumpsd/decode_base.py
@@ -1,0 +1,66 @@
+from collections import OrderedDict
+import struct
+
+class atom(object):
+    '''
+    takes 2-tuple: ('struct_string', 'default_print_format')
+
+    set will set the instance.attribute "val" to the value
+    of the atom's decode of the buffer.
+    '''
+    def __init__(self, a_tuple):
+        self.s_str = a_tuple[0]
+        self.s_rec = struct.Struct(self.s_str)
+        self.p_str = a_tuple[1]
+
+    def __len__(self):
+        return self.s_rec.size
+
+    def __repr__(self):
+        return self.p_str.format(self.val)
+
+    def set(self, buf):
+        '''
+        set the atom.val to the unpack from the format string.
+
+        return the number of bytes (size) consumed
+        '''
+        self.val = self.s_rec.unpack(buf[:self.s_rec.size])[0]
+        return self.s_rec.size
+
+
+class aggie(OrderedDict):
+    '''
+    aggie: aggregation node.
+    takes one parameter a dictionary of key -> {atom | aggie}
+    '''
+    def __init__(self, a_dict):
+        super(aggie, self).__init__(a_dict)
+
+    def __len__(self):
+        l = 0
+        for key, v_obj in self.iteritems():
+            if isinstance(v_obj, atom) or isinstance(v_obj, aggie):
+                l += v_obj.__len__()
+        return l
+
+    def __repr__(self):
+        s = ''
+        for key, v_obj in self.iteritems():
+            if len(s) != 0:
+                s += ', '
+            if isinstance(v_obj, atom):
+                s += key + ': ' + v_obj.__repr__()
+            elif isinstance(v_obj, aggie):
+                s += '[' + key + ': ' + v_obj.__repr__() + ']'
+            else:
+                s += "oops"
+        return s
+
+    def set(self, buf):
+        '''
+        '''
+        consumed = 0
+        for key, v_obj in self.iteritems():
+            consumed += v_obj.set(buf[consumed:])
+        return consumed
