+++ tools/utils/dumpsd/dumpsd/dumpsd.py
@@ -7,6 +7,21 @@ import struct
 import argparse
 from collections import OrderedDict
 
+from decode_base import *
+
+# hdr object at native, little endian
+hdr_obj = aggie(OrderedDict([
+    ('len',  atom(('H', '{}'))),
+    ('type', atom(('H', '{}'))),
+    ('xt',   atom(('I', '0x{:04x}')))]))
+
+
+def print_hdr(obj):
+    rtype = obj['hdr']['type'].val
+    # gratuitous space shows up after the print, sigh
+    print('{:08x} ({:2}) {:6} --'.format(
+        obj['hdr']['xt'].val,
+        rtype, dt_records[rtype][2])),
 
 ####
 #
@@ -35,87 +50,6 @@ DT_H_REVISION       = 0x00000005
 
 LOGICAL_SECTOR_SIZE = 512
 LOGICAL_BLOCK_SIZE  = 508       # excludes trailer
-
-
-class atom(object):
-    '''
-    takes 2-tuple: ('struct_string', 'default_print_format')
-
-    set will set the instance.attribute "val" to the value
-    of the atom's decode of the buffer.
-    '''
-    def __init__(self, a_tuple):
-        self.s_str = a_tuple[0]
-        self.s_rec = struct.Struct(self.s_str)
-        self.p_str = a_tuple[1]
-
-    def __len__(self):
-        return self.s_rec.size
-
-    def __repr__(self):
-        return self.p_str.format(self.val)
-
-    def set(self, buf):
-        '''
-        set the atom.val to the unpack from the format string.
-
-        return the number of bytes (size) consumed
-        '''
-        self.val = self.s_rec.unpack(buf[:self.s_rec.size])[0]
-        return self.s_rec.size
-
-
-class aggie(OrderedDict):
-    '''
-    aggie: aggregation node.
-    takes one parameter a dictionary of key -> {atom | aggie}
-    '''
-    def __init__(self, a_dict):
-        super(aggie, self).__init__(a_dict)
-
-    def __len__(self):
-        l = 0
-        for key, v_obj in self.iteritems():
-            if isinstance(v_obj, atom) or isinstance(v_obj, aggie):
-                l += v_obj.__len__()
-        return l
-
-    def __repr__(self):
-        s = ''
-        for key, v_obj in self.iteritems():
-            if len(s) != 0:
-                s += ', '
-            if isinstance(v_obj, atom):
-                s += key + ': ' + v_obj.__repr__()
-            elif isinstance(v_obj, aggie):
-                s += '[' + key + ': ' + v_obj.__repr__() + ']'
-            else:
-                s += "oops"
-        return s
-
-    def set(self, buf):
-        '''
-        '''
-        consumed = 0
-        for key, v_obj in self.iteritems():
-            consumed += v_obj.set(buf[consumed:])
-        return consumed
-
-
-# hdr object at native, little endian
-hdr_obj = aggie(OrderedDict([
-    ('len',  atom(('H', '{}'))),
-    ('type', atom(('H', '{}'))),
-    ('st',   atom(('Q', '0x{:08x}')))]))
-
-
-def print_hdr(obj):
-    rtype = obj['hdr']['type'].val
-    # gratuitous space shows up after the print, sigh
-    print('{:08x} ({:2}) {:6} --'.format(
-        obj['hdr']['st'].val,
-        rtype, dt_records[rtype][2])),
-
 
 # all dt parts are native and little endian
